// hr-frontend/src/lib/api/generated/model/jobGradeResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface JobGradeResponse {
  gradeId?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  rank?: number;
  active?: boolean;
  positionsCount?: number;
  createdAt?: string;
  updatedAt?: string;
}